requirements.txt
beautifulsoup4==4.9.3
psycopg2-binary
requests==2.25.1